ns: rewrite IApplicationManagerInterface
