+++ src/core/hle/service/am/service/application_functions.cpp
@@ -16,6 +16,7 @@
 #include "core/hle/service/filesystem/filesystem.h"
 #include "core/hle/service/filesystem/save_data_controller.h"
 #include "core/hle/service/glue/glue_manager.h"
+#include "core/hle/service/ns/application_manager_interface.h"
 #include "core/hle/service/ns/ns.h"
 #include "core/hle/service/sm/sm.h"
 
@@ -167,7 +168,7 @@ Result IApplicationFunctions::GetDesiredLanguage(Out<u64> out_language_code) {
     auto app_man = ns_am2->GetApplicationManagerInterface();
 
     // Get desired application language
-    u8 desired_language{};
+    NS::ApplicationLanguage desired_language{};
     R_TRY(app_man->GetApplicationDesiredLanguage(&desired_language, supported_languages));
 
     // Convert to settings language code.
