+++ src/core/hle/service/am/service/library_applet_self_accessor.cpp
@@ -14,6 +14,7 @@
 #include "core/hle/service/cmif_serialization.h"
 #include "core/hle/service/filesystem/filesystem.h"
 #include "core/hle/service/glue/glue_manager.h"
+#include "core/hle/service/ns/application_manager_interface.h"
 #include "core/hle/service/ns/ns.h"
 #include "core/hle/service/sm/sm.h"
 
@@ -260,7 +261,7 @@ Result ILibraryAppletSelfAccessor::GetMainAppletApplicationDesiredLanguage(
     auto app_man = ns_am2->GetApplicationManagerInterface();
 
     // Get desired application language
-    u8 desired_language{};
+    NS::ApplicationLanguage desired_language{};
     R_TRY(app_man->GetApplicationDesiredLanguage(&desired_language, supported_languages));
 
     // Convert to settings language code.
